refactor: move observability fixtures from tests/conftest.py
Extract `prometheus` fixture from the monolithic `tests/conftest.py` into
`tests/fixtures/observability/monitoring.py`, and `audit_logs` fixture
(+ `AUDIT_LOG_PATTERN` constant) into `tests/fixtures/cluster/infrastructure.py`.
Register `tests.fixtures.observability.monitoring` in `pytest_plugins`. Remove
now-unused imports (`subprocess`, `bisect_left`, `defaultdict`, `Prometheus`,
`AUDIT_LOGS_PATH`, `OC_ADM_LOGS_COMMAND`) from `tests/conftest.py`.
